requirements.txt
streamlit
azure-ai-formrecognizer
azure-core
azure-storage-blob
python-dotenv
pandas
Pillow
requests
openpyxl
plotly
httpx
